+++ src/library/model_runner-tools-ooxml.adb
@@ -3,6 +3,8 @@ with Ada.Strings.Unbounded;
 
 with Zlib;
 
+with Model_Runner.Tools.Text_Util;
+
 package body Model_Runner.Tools.OOXML is
 
    package U renames Ada.Strings.Unbounded;
@@ -329,7 +331,8 @@ package body Model_Runner.Tools.OOXML is
          end loop;
       end;
 
-      return U.To_String (Out_Buf);
+      return Model_Runner.Tools.Text_Util.Collapse_Blanks
+        (U.To_String (Out_Buf));
    end Extract_Text;
 
 end Model_Runner.Tools.OOXML;